Threading manager (#5)
* separated threading manager and added functionality to accept planner function
